Add simplified version of the script to run tasks without a workflow manager like rocoto.
The new script (run_one_task_simplified.sh) differs from the original file (run_one_task.sh) in that it does not take as input arguments the job scheduler (sched), the account to which to charge the job (acct), the partition (part) and quality-of-service (qos) to which to submit the job (for slurm scheduler only), and the number of nodes (nnodes), processes per node (ppn), and walltime (wtime).  In the new simplified script, these are read in from the experiment's workflow variable defintions file (var_defns.sh).