Allow valid_param_vals to be machine-specific!
The file ush/valid_param_vals.sh contains arrays of strings that contain all the valid values for a
given user-specified variable. These are checked using the "check_var_valid_value" function, which
currently takes in 2 or 3 arguments: the variable name, the name of the array of valid values, and
an optional error message.

I have modified this script to take an additional optional argument: the machine name. These are
disambiguated using the prefix "msg=" for the error message, and/or "mach=" for the machine name.

The machine-specific arrays build on the previous convention of ush/valid_param_vals.sh: previously
all the arrays of valid values were named "valid_vals_VAR", where VAR was the relevant variable
name for comparison. Now we can *also* allow machine-specific arrays of valid values, using the
naming convention "MACHINE_valid_vals_VAR". If the machine-specific array does not exist for a
given machine, the script falls back to the default non-machine-specific valid_vals_ array.

This effort was undertaken because the valid queue values are different for different machines, so
it is helpful for the user to make this machine-specific to avoid hard-to-solve runtime errors. I
anticipate this capability will be helpful for other variables in the future as well.