This change makes the location of the make_orog fix files more flexible.
Previously, in exregional_make_orog.sh, a local variable "topo_dir" was used as the location of the
fix files for the make_orog task, and this variable was hard-coded as a path relative to the global
"FIXgsm" variable. Now topo_dir is set as a global variable in setup.sh, just like FIXgsm and
SFC_CLIMO_INPUT_DIR, to allow more flexibility for the location of these fixed input files.