Bug fix for background map and removal of yaml
Change from background_img() to imshow() to remove need for a
.json file to show a background map image.

Remove the color palette conversion to allow for background image
and contours to display.

Remove yaml mentions as it is not in use.

Comment out QPF plotting since it is not available the files I am
testing with. Others can comment back in if this is a local error.

Added information in header to point users to the guide for
downloading files needed for Cartopy.