Modify readlink option to not require existence of path.
The -f flag to readlink requires that the last "component" of the path passed to it exist.  In the way readlink is used in ush/setup.sh, the path may not yet exist.  Thus, change the -f to a -m, which states that none of the components of the path need to exist.