Updates to static file structure and naming on Cheyenne. (#399)
## DESCRIPTION OF CHANGES: 
A few updates are needed to the static input and environment setup files for the SRW on Cheyenne. The static files are now located under the `/glade/p/ral/jntp/UFS_SRW_app/` directory rather than the `/glade/p/ral/jntp/UFS_CAM` directory to conform with the latest naming conventions. The python environment has also been renamed to "regional_workflow" to conform with the naming convention of the conda environments on NOAA HPC platforms.

## TESTS CONDUCTED: 
Ran the release suite of end-to-end tests (`tests/testlist.release_public_v1.txt`) and all passed. A few of these were actually previously broken on Cheyenne, and now run to completion successfully.

The only changes here are to Cheyenne-specific code, but I also ran the release suite on Hera as a sanity check (all passed).