Update develop branch for generic linux, MacOS capability (#539)
## DESCRIPTION OF CHANGES: 
This change will add the capability to run regional_workflow (as part of the SRW app) on MacOS and generic LINUX platforms. Most of these changes are identical to those in #402 (hash bc08607) but some additional modifications needed to be made due to intervening changes in the develop branch.

## TESTS CONDUCTED: 
Ran Graduate Student Test on new platforms:
 - my personal Mac machine (MacOS Catalina 10.15.7) MacOS with gnu 9.4.0 compilers. 
 - Cheyenne compute node as a faux "stand-alone" machine, intel 19.1.1 compilers

Ran suite of end-to-end tests on Cheyenne (intel/19.1.1) and Hera (intel/18.0.5.274). All passed as expected.

Tests also passed on WCOSS, MacOS Mojave, RedHat Linux.

## ISSUE: 
Will resolve #369