Reduce the default wallclock time for the make_orog task.  Details below.
Reduce WTIME_MAKE_OROG from 1 hour to 20 minutes since on Hera, 1 hour causes the task to wait quite a while in the queue.  This used to be 20 minutes but was changed to 1 hour because 20 minutes was not sufficient to complete the task on WCOSS_DELL_P3.  Thus, reset it to 1 hour only in the machine file for WCOSS_DELL_P3.