Fix the way wallclock times are set to prevent conversion to seconds in the ROCOTO XML.  Details below.
In the yaml-formatted variable "settings" that contains the jinja settings to be used with the XML template FV3LAM_wflow.xml, place all wall clock time variables (${WTIME_...}) in single quotes.  If this is not done, wall clock times set to 10:00:00 (i.e. 10 hours) and possibly other values first get converted to seconds (36000) before being written to the XML, which causes an error when processed by ROCOTO.  (The reason for this unit conversion is unclear.)